% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_base_data.R
\name{load_base_data}
\alias{load_base_data}
\title{Load base unit data .json files into a tibble}
\usage{
load_base_data(base_dir = file.path("inst", "units", "base"))
}
\arguments{
\item{base_dir}{A file pathway where base unit .json files are stored.
Defaults to package data directory file.path("inst", "units", "base")}
}
\description{
Get all base unit data from .json files into a tibble for onward
manipulation.
}
\keyword{internal}
