% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_derived_data.R
\name{load_derived_data}
\alias{load_derived_data}
\title{Load derived unit data .json files into a tibble}
\usage{
load_derived_data(derived_dir = file.path("inst", "units", "derived"))
}
\arguments{
\item{derived_dir}{A file pathway where derived unit .json files are stored.
Defaults to package data directory file.path("inst", "units", "derived")}
}
\description{
Get all derived unit data from .json files into a tibble for onward
manipulation.
}
\keyword{internal}
