% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_json_files.R
\name{load_json_files}
\alias{load_json_files}
\title{Load .json files}
\usage{
load_json_files(file_pathway)
}
\arguments{
\item{file_pathway}{A file pathway where .json files are stored.
E.g. file.path("inst", "units", "base")}
}
\description{
Small helper function to load all .json files within a directory
and its sub folders.
}
\keyword{internal}
