% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_own_unicon_data.R
\name{make_own_unicon_data}
\alias{make_own_unicon_data}
\title{Make package data using user's own addins}
\usage{
make_own_unicon_data(user_base_dir = NA, user_derived_dir = NA)
}
\arguments{
\item{user_base_dir}{A file pathway where base unit .json files are stored.
Defaults to package data directory file.path("inst", "units", "base")}

\item{user_derived_dir}{A file pathway where derived unit .json files are stored.
Defaults to package data directory file.path("inst", "units", "derived")}
}
\description{
Calculate necessary dataframes used in other package functions
using user's own additional data. If using package's own data not necessary.
}
