% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon.full.R
\name{unicon.full}
\alias{unicon.full}
\title{Universal unit conversion}
\usage{
unicon.full(value.in, unit.in, unit.out = NA, pull = TRUE)
}
\arguments{
\item{value.in}{Numeric scalar or vector, values to convert.}

\item{unit.in}{Character scalar or vector, input units for \code{value.in}. Must be of
\code{length(1L)} or \code{length(value.in)}.}

\item{unit.out}{Character scalar or vector, output units for conversion. Must be of
\code{length(1L)} or \code{length(value.in)}. Defaults to \code{NA}; if default is passed,
function will return standard index (SI) units as conversion.}

\item{pull}{Logical; should the function pull out and return the converted values (TRUE) or
should a full table with conversion record be returned? Defaults to TRUE.}
}
\description{
Full-service/user-exposable unit conversion. Supply a vector of input values with
associated unit aliases and receive conversions to any required output alias. Designed to handle
the widest possible range of naming conventions, permutations and plain mis-spellings of commonly
used units.
}
