% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon.lite.R
\name{unicon.lite}
\alias{unicon.lite}
\title{Universal unit conversion}
\usage{
unicon.lite(value.in, id.in, id.out = NA, pull = TRUE)
}
\arguments{
\item{value.in}{Numeric scalar or vector, values to convert.}

\item{id.in}{Character scalar or vector, input unit ID(s) for \code{value.in}. Must be of
\code{length(1L)} or \code{length(value.in)}.}

\item{id.out}{Character scalar or vector, output unit ID(s) for conversion. Must be of
\code{length(1L)} or \code{length(value.in)}. Defaults to \code{NA}; if default is passed,
function will return standard index (SI) units as conversion.}

\item{pull}{Logical; should the function pull out and return the converted values (TRUE) or
should a full table with conversion record be returned? Defaults to TRUE.}
}
\description{
Lite/internal use function for unit conversion; supply a vector of input values with
associated unit IDs and receive conversions to any required outputs.
}
