% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_advance.R
\name{unicon_advance}
\alias{unicon_advance}
\title{Deriving and calculating unit conversion between two metrics}
\usage{
unicon_advance(
  x_value_in,
  y_value_in,
  x_unit_in,
  y_unit_in,
  unit_out = NA,
  operator_in = NA,
  pull = TRUE
)
}
\arguments{
\item{x_value_in}{Numeric scalar or vector, values to convert from \code{x} and
use in calculations of new metrics}

\item{y_value_in}{Numeric scalar or vector, values to convert from \code{y} and
use in calculations of new metric}

\item{x_unit_in}{Character scalar or vector, input units for
\code{x_value_in}. Must be of \code{length(1L)} or \code{length(x_value_in)}.}

\item{y_unit_in}{Character scalar or vector, input units for
\code{y_value_in}. Must be of \code{length(1L)} or \code{length(y_value_in)}.}

\item{unit_out}{Character scalar or vector, output units for conversion. Must
be of \code{length(1L)} or \code{length(x_value_in)}. Defaults to
\code{NA}; if default is passed, function will return standard reference
point (SRP) units as conversion.}

\item{operator_in}{Character scalar; \code{'divide'} or \code{'multiply'}.
Only needed for calculating volume from volume_fraction and mass from
mass_fraction.}

\item{pull}{Logical; should the function pull out and return the converted
values (TRUE) or should a full table with conversion record be returned?
Defaults to TRUE.}
}
\description{
Using the \code{relationships} package data, this function uses
'unicon_full' to calculate new metrics from parent metrics by firstly
converting parent units to standard reference point (SRP), calculating the
new metric in SRP units then converting to the desired units if specified.
}
