% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_full.R
\name{unicon_full}
\alias{unicon_full}
\title{Universal unit conversion}
\usage{
unicon_full(value_in, unit_in, unit_out = NA, pull = TRUE)
}
\arguments{
\item{value_in}{Numeric scalar or vector, values to convert.}

\item{unit_in}{Character scalar or vector, input units for \code{value_in}.
Must be of \code{length(1L)} or \code{length(value_in)}.}

\item{unit_out}{Character scalar or vector, output units for conversion. Must
be of \code{length(1L)} or \code{length(value_in)}. Defaults to \code{NA}; if
default is passed, function will return standard reference point (SRP) (which
is a standard index (SI) unit) conversion.}

\item{pull}{Logical; should the function pull out and return the converted
values (TRUE) or should a full table with conversion record be returned?
Defaults to TRUE.}
}
\description{
Full-service/user-exposable unit conversion. Supply a vector of
input values with associated unit aliases and receive conversions to any
required output alias. Designed to handle the widest possible range of naming
conventions, permutations and plain mis-spellings of commonly used units.
This is a wrapper for unicon_lite where just looks up the alias.
}
