% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_help.R
\name{unicon_help}
\alias{unicon_help}
\title{Helper for unit conversion}
\usage{
unicon_help(...)
}
\arguments{
\item{...}{Optional named character scalar arguments used as regular
expressions to filter the output table by one or more columns (e.g.
\code{category = "mass"}). May be omitted if the full table is required.}
}
\description{
Return full unit conversion table with optional filter-by-regex
args.
}
