% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_join_datasets.R
\name{unicon_join_datasets}
\alias{unicon_join_datasets}
\title{Join .json tibbles together}
\usage{
unicon_join_datasets(base, derived, operators)
}
\arguments{
\item{base}{A tibble with the columns "id", "alias", "category", "srp",
"slope", "intercept"}

\item{derived}{A tibble with the columns "id", "x", "y", "operator"}

\item{operators}{A tibble with the columns "operator", "id", "fun", "alias"}
}
\value{
A dataframe with all information from .jsons inside mapped at all
levels
}
\description{
Join base, derived, operator datasets together so that every
combination has been mapped.
}
