% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_data_from_jsons.R
\name{unicon_load_json_files}
\alias{unicon_load_json_files}
\title{Load .json files}
\usage{
unicon_load_json_files(file_pathway)
}
\arguments{
\item{file_pathway}{The file pathway where .json files are stored. Note this
is the top level 'parent' folder; all sub folder wll also be examined for
.json files}
}
\value{
A dataframe with all information from .jsons inside
}
\description{
Generic helper function for loading json files
}
