% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_data_from_jsons.R
\name{unicon_make_base_data_from_jsons}
\alias{unicon_make_base_data_from_jsons}
\title{Load base data .jsons}
\usage{
unicon_make_base_data_from_jsons(dir)
}
\arguments{
\item{dir}{The file pathway where .json files for base data are
stored. Please refer to vignette for more information.}
}
\description{
Using the \code{load_json_files} wrapper function, makes a tibble
with the columns "id", "alias", "category", "srp", "slope", "intercept" that
were formally lists.
}
