% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_data_from_jsons.R
\name{unicon_make_operators_data_from_jsons}
\alias{unicon_make_operators_data_from_jsons}
\title{Load operator data .jsons}
\usage{
unicon_make_operators_data_from_jsons(dir)
}
\arguments{
\item{dir}{The file pathway where .json files for operator data are
stored. Please refer to vignette for more information.}
}
\description{
Using the \code{load_json_files} wrapper function, makes a tibble
with the columns "operator", "id", "fun", "alias" that were formally lists.
}
