% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_own_data.R
\name{unicon_make_own_derived_data}
\alias{unicon_make_own_derived_data}
\title{Make your own derived data for use with unicon}
\usage{
unicon_make_own_derived_data(id, x, y, operator)
}
\arguments{
\item{id}{Character scalar or vector; the new derived metric}

\item{x}{Character scalar or vector; the base (numerator) metric that is
used to calculate the current metric}

\item{y}{Character scalar or vector; the base (denominator) metric that is
used to calculate the current metric}

\item{operator}{Character scalar or vector; the operator used to calculate
the metric using x and y}
}
\value{
A tibble with columns \code{id}, \code{x}, \code{y}, and
\code{operator}.
}
\description{
If a user wants to use unicon but the derived data they want is not in the
package data, they are welcome to add their own. This function is for the
derived data i.e. if two measures are calculated to make another measure.
When designing your own data please reference the data available using
\code{unicon_help}. This is used at the users own discretion and they will need to
conduct their own checks for data quality.
}
