% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_relationships_data.R
\name{unicon_make_relationships_data}
\alias{unicon_make_relationships_data}
\title{Category relationships}
\usage{
unicon_make_relationships_data(derived)
}
\arguments{
\item{derived}{A dataframe made from \code{unicon_make_derived_data_from_jsons}}
}
\value{
Makes a dataframe with all combination of relationships between
unit categories, where any unit category can be every combination of \code{id},
\code{x}, and \code{y} with the correct \code{operator} assigned.
}
\description{
Category relationships
}
