% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_unit.R
\name{unicon_make_unit_alias}
\alias{unicon_make_unit_alias}
\title{Make unit_alias dataset}
\usage{
unicon_make_unit_alias(join_dataset)
}
\arguments{
\item{join_dataset}{The tibble created by the \code{unicon_join_datasets} function}
}
\value{
A dataframe with all "id" and "alias" combinations
}
\description{
All unique aliases with matching id
}
