% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_unit.R
\name{unicon_make_unit_models}
\alias{unicon_make_unit_models}
\title{Make unit_model dataset}
\usage{
unicon_make_unit_models(join_dataset)
}
\arguments{
\item{join_dataset}{The tibble created by the \code{unicon_join_datasets} function}
}
\value{
A dataframe with all "id" and "model" combinations, where "model" is
a list of "slope" and "intercept"
}
\description{
All unique models
}
