% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_unit.R
\name{unicon_make_unit_srp}
\alias{unicon_make_unit_srp}
\title{Make unit_srp dataset}
\usage{
unicon_make_unit_srp(join_dataset)
}
\arguments{
\item{join_dataset}{The tibble created by the \code{unicon_join_datasets} function}
}
\value{
A dataframe with all "id", "type", "category" and "srp" combinations
}
\description{
All unique standard reference points
}
