% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_option.R
\name{unicon_option}
\alias{unicon_option}
\title{Deriving and calculating unit conversion between one or two metrics}
\usage{
unicon_option(value_in, unit_in, unit_out, extras)
}
\arguments{
\item{value_in}{Numeric scalar or vector, values to convert.}

\item{unit_in}{Character scalar or vector, input units for \code{value_in}.
Must be of \code{length(1L)} or \code{length(value_in)}.}

\item{unit_out}{Character scalar or vector, output units for conversion. Must
be of \code{length(1L)} or \code{length(value_in)}.}

\item{extras}{A list of other unit categories needed for use in
\code{unicon_advance}, which themsevles are a list of two specificying \code{unit} and
\code{value}. Must use unit category names (see \code{unicon_help()}).}
}
\description{
When you have a mixture of unit categories, but want one unit
outcome. This function uses both unicon_full and unicon_advance to give a
single value outcome in a specified unit.
}
