% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{unicon_own_status}
\alias{unicon_own_status}
\title{Tells user if using their own data or package data. \code{TRUE} is using
user's own data.}
\usage{
unicon_own_status()
}
\description{
Tells user if using their own data or package data. \code{TRUE} is using
user's own data.
}
