% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conv.R
\name{unit_conv}
\alias{unit_conv}
\title{unit_conv converts measurement data (e.g. km, gallons, kelvin) into Trinity's standard units.
Input \emph{must} be a dataframe.
To apply to multiple columns please use \code{unit_conv_scale}.}
\usage{
unit_conv(df, measure, quantity)
}
\arguments{
\item{df}{a data frame}

\item{measure}{the column that contains the units - must be character}

\item{quantity}{the column that contains the values - must be integer}
}
\value{
a dataframe with two additional columns, si and si_quantity
}
\description{
unit_conv converts measurement data (e.g. km, gallons, kelvin) into Trinity's standard units.
Input \emph{must} be a dataframe.
To apply to multiple columns please use \code{unit_conv_scale}.
}
\examples{
d = data.frame(unit = c('inch',
                        'cm',
                        'meters',
                        'mile',
                        'yard',
                        'stone',
                        'pints',
                        'mls',
                        'k',
                        'uk_gallon'),
               value = c(10:19))

d_si = d |>
  unit_conv(measure = unit,
            quantity = value)

Warning message:
In unit_conv(d, measure = unit, quantity = value) :
  NA values detected — results may be incomplete
}
