% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_base_json.R
\name{write_base_json}
\alias{write_base_json}
\title{Write base unit .json files}
\usage{
write_base_json(category, si, slope, intercept = 0, alias, path)
}
\arguments{
\item{category}{A character vector of L1. The type of measurement the unit is.}

\item{si}{A character vector of L1. The SI unit that the package will
automatically convert to.}

\item{slope}{A numeric vector of L1. The ratio between the unit in relation to the SI unit.}

\item{intercept}{A numeric vector of L1. Defaults to 0. Only use if relationship
betwen created metric and SI starts at different points e.g. celcius and kelvin.}

\item{alias}{A character vector. A list of names used to describe the unit being added.}

\item{path}{A file pathway and file name. Note file name will be the unit id.}
}
\description{
If a user wishes to add to the package's data using their own,
they need to create .json files with the relevant information for base units.
}
