% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_derived_json.R
\name{write_derived_json}
\alias{write_derived_json}
\title{Write derived unit .json files}
\usage{
write_derived_json(x, y, operator, path)
}
\arguments{
\item{x}{A character vector of L1. The numerator in fraction.}

\item{y}{A character vector of L1. The denominator in fraction.}

\item{operator}{A numeric vector of L1. Either 'divide' or 'multiply'.}

\item{path}{A file pathway and file name. Note file name will be the unit id.}
}
\description{
If a user wishes to add to the package's data using their own,
they need to create .json files with the relevant information for derived units,
assuming the relationship is either a fraction where x is numerator and y is
denominator or a multiplication.
}
