% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_help.R
\name{unicon_help}
\alias{unicon_help}
\title{Helper for unit conversion}
\usage{
unicon_help(...)
}
\arguments{
\item{...}{Optional filter parameters, named regular expressions for
filtering by one or more variables. May be ignored if full table is required.}
}
\description{
Return full unit conversion table with optional filter-by-regex
args.
}
