% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_lite.R
\name{unicon_lite}
\alias{unicon_lite}
\title{Universal unit conversion}
\usage{
unicon_lite(value_in, id_in, id_out = NA)
}
\arguments{
\item{value_in}{Numeric scalar or vector, values to convert.}

\item{id_in}{Character scalar or vector, input unit ID(s) for
\code{value_in}. Must be of \code{length(1L)} or \code{length(value_in)}.}

\item{id_out}{Character scalar or vector, output unit ID(s) for conversion.
Must be of \code{length(1L)} or \code{length(value_in)}. Defaults to
\code{NA}; if default is passed, function will return the standard
reference point (SRP), which is a standard index (SI) unit measure.}
}
\description{
The function that does the actual unit conversion and heavy
lifting. To only be used as part of unicon_full to ensure data checks
completed.
}
