% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_lite.R
\name{unicon_lite}
\alias{unicon_lite}
\title{Universal unit conversion}
\usage{
unicon_lite(value_in, id_in, id_out = NA, pull = TRUE)
}
\arguments{
\item{value_in}{Numeric scalar or vector, values to convert.}

\item{id_in}{Character scalar or vector, input unit ID(s) for
\code{value_in}. Must be of \code{length(1L)} or \code{length(value_in)}.}

\item{id_out}{Character scalar or vector, output unit ID(s) for conversion.
Must be of \code{length(1L)} or \code{length(value_in)}. Defaults to
\code{NA}; if default is passed, function will return standard index (SI)
units as conversion.}

\item{pull}{Logical; should the function pull out and return the converted
values (TRUE) or should a full table with conversion record be returned?
Defaults to TRUE.}
}
\description{
Lite/internal use function for unit conversion; supply a vector
of input values with associated unit IDs and receive conversions to any
required outputs.
}
