% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_own_data.R
\name{unicon_make_own_base_data}
\alias{unicon_make_own_base_data}
\title{make your own base data for use with unicon's functionality}
\usage{
unicon_make_own_base_data(id, alias, category, srp, slope, intercept = 0)
}
\arguments{
\item{id}{the default name of the metric, needs to be consistent across
aliases}

\item{alias}{alternative names for the metric, many of these can map to id}

\item{category}{what is the unit measuring?}

\item{srp}{the standard reference point, the default unit of the category}

\item{slope}{the difference between the unit and srp unit}

\item{intercept}{defaults to 0. Only use if relationship is not linear}
}
\value{
A tibble with columns \code{id}, \code{alias}, \code{category},
\code{srp}, \code{slope} and \code{intercept}.
}
\description{
If a user wants to use unicon but the unit they want is not in the package
data, they are welcome to add their own. This function is for the base data
i.e. units cannot be derived from other units. When designing your own data
please reference the data available using \code{unicon_help}. This is used at the
users own discretion and they will need to conduct their own checks for data
quality.
}
