% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicon_make_own_data.R
\name{unicon_make_own_derived_data}
\alias{unicon_make_own_derived_data}
\title{If a user wants to use unicon but the derived data they want is not in the
package data, they are welcome to add their own. This function is for the
derived data i.e. if two measures are calculated to make another measure.
When designing your own data please reference the data available using
\code{unicon_help}. This is used at the users own discretion and they will need to
conduct their own checks for data quality.}
\usage{
unicon_make_own_derived_data(id, x, y, operator)
}
\arguments{
\item{id}{the new derived metric}

\item{x}{the base (numerator) metric that is used to calculate the current
metric}

\item{y}{the base (denominator) metric that is used to calculate the current
metric}

\item{operator}{the operator used to calculate the metric using x and y}
}
\description{
If a user wants to use unicon but the derived data they want is not in the
package data, they are welcome to add their own. This function is for the
derived data i.e. if two measures are calculated to make another measure.
When designing your own data please reference the data available using
\code{unicon_help}. This is used at the users own discretion and they will need to
conduct their own checks for data quality.
}
