% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{unicon_reset_units}
\alias{unicon_reset_units}
\title{Reset the \verb{unit_*} datasets to the package stored ones.
Used when package loads or to revert after using custom data.}
\usage{
unicon_reset_units()
}
\description{
Reset the \verb{unit_*} datasets to the package stored ones.
Used when package loads or to revert after using custom data.
}
